+++ lib/ruby2js/converter/send.rb
@@ -303,6 +303,15 @@ module Ruby2JS
       elsif method == :typeof and receiver == nil
         put 'typeof '; parse args.first
 
+      elsif ast.children[1] == :is_a? and receiver and args.length == 1
+        parse receiver; put ' instanceof '; parse args.first
+
+      elsif ast.children[1] == :kind_of? and receiver and args.length == 1
+        parse receiver; put ' instanceof '; parse args.first
+
+      elsif ast.children[1] == :instance_of? and receiver and args.length == 1
+        parse s(:send, s(:attr, receiver, :constructor), :==, args.first)
+
       else
         put 'await ' if @ast.type == :await